#11 Get corner cutter config from backend
