+++ frontend/src/redux/slices/cornercutter/index.ts
@@ -1,0 +1,2 @@
+export { default } from './cornercutter.slice';
+export * from './cornercutter.slice';
