+++ frontend/src/types/CornerCutterConfig.ts
@@ -1,0 +1,3 @@
+export interface CornerCutterConfig {
+    goingUnderDir?: string;
+}
